app/requirements.txt
flask
gunicorn
psycopg2-binary
prometheus-client